release: remove obsolete browser continuity tooling
