+++ .github/workflows/docker-publish.yml
@@ -259,12 +259,6 @@ jobs:
           ref: ${{ needs.validate.outputs.release_commit }}
           persist-credentials: false
 
-      - name: Check out release tooling
-        uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1 # v7.0.1
-        with:
-          path: .release-source
-          persist-credentials: false
-
       - name: Download and verify exact release archive
         env:
           GH_TOKEN: ${{ github.token }}
@@ -280,10 +274,7 @@ jobs:
           printf '%s  %s\n' "${ASSET_DIGEST#sha256:}" "$archive" | sha256sum -c -
           tar -tzf "$archive" >/dev/null
           cp ci/release/docker/entrypoint.sh "$RUNNER_TEMP/docker-context/entrypoint.sh"
-          bash .release-source/ci/release/docker/prepare-continuity-dockerfile.sh \
-            "$VERSION" \
-            ci/release/docker/Dockerfile \
-            "$RUNNER_TEMP/Dockerfile"
+          cp ci/release/docker/Dockerfile "$RUNNER_TEMP/Dockerfile"
 
       - name: Log in to Docker Hub
         env:
@@ -391,12 +382,6 @@ jobs:
           ref: ${{ needs.validate.outputs.release_commit }}
           persist-credentials: false
 
-      - name: Check out release tooling
-        uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1 # v7.0.1
-        with:
-          path: .release-source
-          persist-credentials: false
-
       - name: Download and verify exact release archive
         env:
           GH_TOKEN: ${{ github.token }}
@@ -412,10 +397,7 @@ jobs:
           printf '%s  %s\n' "${ASSET_DIGEST#sha256:}" "$archive" | sha256sum -c -
           tar -tzf "$archive" >/dev/null
           cp ci/release/docker/entrypoint.sh "$RUNNER_TEMP/docker-context/entrypoint.sh"
-          bash .release-source/ci/release/docker/prepare-continuity-dockerfile.sh \
-            "$VERSION" \
-            ci/release/docker/Dockerfile \
-            "$RUNNER_TEMP/Dockerfile"
+          cp ci/release/docker/Dockerfile "$RUNNER_TEMP/Dockerfile"
 
       - name: Log in to Docker Hub
         env:
